django-backend/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
Django==5.0.7
django-cleanup==8.1.0
django-cors-headers==4.4.0
django-ninja==1.2.2
dnspython==2.6.1
email_validator==2.2.0
idna==3.7
pillow==10.4.0
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.8.0
python-dotenv==1.0.1
sqlparse==0.5.1
typing_extensions==4.12.2
